# contracts/openapi/dashboard-v1.json
{
  "components": {
    "schemas": {
...
      "AnalysisSessionResponse": {
        "additionalProperties": false,
        "properties": {
          "bytes_downloaded": {
            "minimum": 0.0,
            "title": "Bytes Downloaded",
            "type": "integer"
          },
          "cancellation_requested": {
            "title": "Cancellation Requested",
            "type": "boolean"
          },
          "candidate_count": {
            "maximum": 30.0,
            "minimum": 0.0,
            "title": "Candidate Count",
            "type": "integer"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/AnalysisCandidateResponse"
            },
            "title": "Candidates",
            "type": "array"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "document_types": {
            "items": {
              "$ref": "#/components/schemas/DocumentType"
            },
            "title": "Document Types",
            "type": "array"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "firecrawl_connection_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firecrawl Connection Id"
          },
          "firecrawl_connection_name_snapshot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firecrawl Connection Name Snapshot"
          },
          "firecrawl_connection_type_snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConnectionType"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "job_stage": {
            "$ref": "#/components/schemas/DiscoveryJobStage"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "session_byte_limit": {
            "minimum": 1.0,
            "title": "Session Byte Limit",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisSessionStatus"
          },
          "tables_required": {
            "title": "Tables Required",
            "type": "boolean"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "query",
          "document_types",
          "tables_required",
          "firecrawl_connection_id",
          "firecrawl_connection_name_snapshot",
          "firecrawl_connection_type_snapshot",
          "status",
          "job_stage",
          "error_code",
          "candidate_count",
          "bytes_downloaded",
          "session_byte_limit",
          "cancellation_requested",
          "created_at",
          "updated_at",
          "expires_at",
          "completed_at",
          "candidates"
        ],
        "title": "AnalysisSessionResponse",
        "type": "object"
      },
      "AnalysisSessionStatus": {
        "enum": [
          "queued",
          "running",
          "completed",
          "cancelled",
          "failed"
        ],
        "title": "AnalysisSessionStatus",
        "type": "string"
      },
...
      "ConnectionType": {
        "enum": [
          "cloud",
          "remote"
        ],
        "title": "ConnectionType",
        "type": "string"
      },
...
      "DiscoveryJobStage": {
        "enum": [
          "queued",
          "discovering",
          "analyzing",
          "completed",
          "cancelled",
          "failed"
        ],
        "title": "DiscoveryJobStage",
        "type": "string"
      },
      "DocumentCandidate": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "document_type": {
            "$ref": "#/components/schemas/DocumentType"
          },
          "source": {
            "const": "firecrawl",
            "default": "firecrawl",
            "title": "Source",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "url": {
            "format": "uri",
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "document_type"
        ],
        "title": "DocumentCandidate",
        "type": "object"
      },
      "DocumentDiscoveryRequest": {
        "additionalProperties": false,
        "properties": {
          "document_types": {
            "default": [
              "pdf",
              "docx"
            ],
            "items": {
              "$ref": "#/components/schemas/DocumentType"
            },
            "title": "Document Types",
            "type": "array"
          },
          "exclude_domains": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Exclude Domains",
            "type": "array"
          },
          "firecrawl_connection_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firecrawl Connection Id"
          },
          "include_domains": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Include Domains",
            "type": "array"
          },
          "limit": {
            "default": 20,
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "query": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "DocumentDiscoveryRequest",
        "type": "object"
      },
      "DocumentDiscoveryResponse": {
        "additionalProperties": false,
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/DocumentCandidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "provider_search_ids": {
            "items": {
              "type": "string"
            },
            "title": "Provider Search Ids",
            "type": "array"
          },
          "rejected_non_document_results": {
            "title": "Rejected Non Document Results",
            "type": "integer"
          }
        },
        "required": [
          "provider_search_ids",
          "candidates",
          "rejected_non_document_results"
        ],
        "title": "DocumentDiscoveryResponse",
        "type": "object"
      },
...
      "DocumentType": {
        "enum": [
          "pdf",
          "docx"
        ],
        "title": "DocumentType",
        "type": "string"
      },
      "DurableAnalysisSearchRequest": {
        "additionalProperties": false,
        "properties": {
          "document_types": {
            "default": [
              "pdf",
              "docx"
            ],
            "items": {
              "$ref": "#/components/schemas/DocumentType"
            },
            "title": "Document Types",
            "type": "array"
          },
          "exclude_domains": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Exclude Domains",
            "type": "array"
          },
          "firecrawl_connection_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firecrawl Connection Id"
          },
          "force_refresh": {
            "default": false,
            "title": "Force Refresh",
            "type": "boolean"
          },
          "include_domains": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Include Domains",
            "type": "array"
          },
          "limit": {
            "default": 20,
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "query": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "tables_required": {
            "default": true,
            "title": "Tables Required",
            "type": "boolean"
          }
        },
        "required": [
          "query"
        ],
        "title": "DurableAnalysisSearchRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Parserium Collector",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/discovery/analyses/{candidate_id}/preview": {
      "get": {
        "operationId": "get_analysis_preview",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "parserium_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parserium Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Analysis Preview",
        "tags": [
          "analysis"
        ]
      }
    },
    "/api/v1/discovery/search": {
      "post": {
        "operationId": "search_documents",
        "parameters": [
          {
            "in": "header",
            "name": "X-Parserium-CSRF",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Parserium-Csrf"
            }
          },
          {
            "in": "cookie",
            "name": "parserium_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parserium Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentDiscoveryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDiscoveryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Documents",
        "tags": [
          "discovery"
        ]
      }
    },
    "/api/v1/discovery/searches": {
      "post": {
        "operationId": "create_analysis_search",
        "parameters": [
          {
            "in": "header",
            "name": "X-Parserium-CSRF",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Parserium-Csrf"
            }
          },
          {
            "in": "cookie",
            "name": "parserium_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parserium Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DurableAnalysisSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Analysis Search",
        "tags": [
          "analysis"
        ]
      }
    },
    "/api/v1/discovery/searches/{session_id}": {
      "delete": {
        "operationId": "cancel_analysis_search",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Parserium-CSRF",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Parserium-Csrf"
            }
          },
          {
            "in": "cookie",
            "name": "parserium_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parserium Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Analysis Search",
        "tags": [
          "analysis"
        ]
      },
      "get": {
        "operationId": "get_analysis_search",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "parserium_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parserium Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Analysis Search",
        "tags": [
          "analysis"
        ]
      }
    },
    "/api/v1/discovery/searches/{session_id}/retry": {
      "post": {
        "operationId": "retry_analysis_search",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Parserium-CSRF",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Parserium-Csrf"
            }
          },
          {
            "in": "cookie",
            "name": "parserium_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parserium Session"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Analysis Search",
        "tags": [
          "analysis"
        ]
      }
    },
...
  }
}